`consistent-boolean-name`: Refine `booleanWrappers` support
Fixes #3556
