+++ docs/rules/consistent-boolean-name.md
@@ -15,7 +15,7 @@ Boolean names should start with a prefix that makes the boolean meaning clear.
 
 Names that start with a boolean prefix should also refer to booleans or boolean-returning functions. Unknown values are ignored.
 
-Configured wrapper bindings may use boolean prefixes when their configured value member provides a boolean value. The wrapper itself is not treated as a boolean for other checks. Only variable and parameter bindings are supported; object and class properties are not affected. The binding must not be reassigned after initialization.
+Configured wrapper bindings may use boolean prefixes when their configured value member provides a boolean-like value. The wrapper itself is not treated as a boolean for other checks. Only variable and parameter bindings are supported; object and class properties are not affected. The binding must not be reassigned after initialization.
 
 Reports for property and method names, and reports for non-boolean values using boolean prefixes, do not provide rename suggestions.
 
@@ -238,7 +238,7 @@ const didUpdate = true;
 Type: `Record<string, string>`\
 Default: `{}`
 
-Allow configured wrapper bindings to use boolean prefixes when a property or method provides a boolean value. The object key is the unqualified TypeScript wrapper type name, and the value is the property or method name. Derived types, intersections, and type parameters constrained by a configured wrapper type are supported. The member may provide `boolean`, `Promise<boolean>`, or `PromiseLike<boolean>`; nullable boolean results are also accepted. This option requires TypeScript type information.
+Allow configured wrapper bindings to use boolean prefixes when a property or method provides a boolean-like value. The object key is the unqualified TypeScript wrapper type name, and the value is the property or method name. Derived types, intersections, and type parameters constrained by a configured wrapper type are supported. The member may provide `boolean`, `Promise<boolean>`, or `PromiseLike<boolean>`; nullable boolean results are also accepted. This option requires TypeScript type information.
 
 ```js
 'unicorn/consistent-boolean-name': [
